+++ arduino/CAR_PC/useful.h
@@ -7,7 +7,7 @@ class USEFUL
   public:
     static unsigned int hexToDec(String hexString);
     static void freeMode();
-    static String convertSecToHMstring(int sec);
+    static String convertSecToHMstring(long sec);
 };
 #endif
